pyproject.toml
[tool.black]
line-length = 140
target-version = ['py311']
include = '\.pyi?$'

[tool.poetry]
name = "mqtt2influxdb"
version = "0.0.1"
description = ""
authors = [
	"Tobias Müller <Tobias_Mueller@twam.info>"
]
license = "MIT"

[tool.poetry.scripts]
mqtt2influxdb = 'mqtt2influxdb.mqtt2influxdb:main'

[tool.poetry.dependencies]
python="~3.11"
python-daemon="~3.0.1"
PyYAML="~6.0"
paho-mqtt="~1.6"
influxdb="*"

[tool.poetry.dev-dependencies]
black = "~23.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
